Fix build when no rtti is available.
This is much needed by the HTML5 platform to decrease WASM size, and
getting better performances on the web.
Additionally, this is only used for a print verbose, so it's not a great
loss.

Now the question is, is the standard macro good enough? Or should I make
a NO_RTTI define specifically for JS?