[HTML5] HTTP server uses optional SSL.
Generates a key/cert snakeoil pair or use a custom SSL cert/key.
This is of course false security, and potentially dangerous.

But, so long, those are the requirements browser vendors agree on to try out
things as simple as the Gamepad API, and more advanced topics like wasm
threads. You don't need this if you run on localhost (at least!), but you do
need thise (or a much safer nginx proxy) to try those things on your local
network (e.g. when debugging a phone, networking, etc).

Oh the web is a very funny place, sir!